Create Immutable class, delete unused imports and methods, fix json write and read bug

+++ tic-tac-toe/tic-tac-toe/MenuForm.cs
@@ -3,23 +3,27 @@ using System.Collections.Generic;
 using System.Windows.Forms;
 using System.IO;
 using System.Text.Json;
+using System.Collections.Immutable;
 
 namespace tic_tac_toe
 {
+
     public partial class MenuForm : Form
     {
 
         public static List<Player> players = new List<Player>();
+        public static GameConfig gameConfig = new GameConfig("tic-tac-toe");
 
         public MenuForm()
         {
             InitializeComponent();
             loadPlayersData();
+            this.Text = gameConfig.GameName;
         }
 
         public void loadform(object Form)
         {
-            if(this.mainpanel.Controls.Count > 0)
+            if (this.mainpanel.Controls.Count > 0)
             {
                 this.mainpanel.Controls.RemoveAt(0);
             }
@@ -31,31 +35,21 @@ namespace tic_tac_toe
             f.Show();
         }
 
-        public static void savePlayersData()
-        {
-            List<Player> playerList = new List<Player>();
-            playerList.Add(new Player()
-            {
-                Name = "Teszt1",
-                Loses = 12,
-                Wins = 30
-            });
-            playerList.Add(new Player()
-            {
-                Name = "Teszt2",
-                Loses = 14,
-                Wins = 55
-            });
-            string playersJson = JsonSerializer.Serialize(playerList);
-            File.WriteAllText(@"C:\Users\gabri\Desktop\ISKOLA\6.félév\C#\egyetemikurzus-2022\toplist.json", playersJson);
-        }
-
         public static void loadPlayersData()
         {
-            using (StreamReader r = new StreamReader(@"C:\Users\gabri\Desktop\ISKOLA\6.félév\C#\egyetemikurzus-2022\toplist.json"))
+            if (File.Exists(@"C:\Users\gabri\Desktop\ISKOLA\6.félév\C#\egyetemikurzus-2022\toplist.json"))
             {
-                string json = r.ReadToEnd();
-                players = JsonSerializer.Deserialize<List<Player>>(json);
+                using (StreamReader r = new StreamReader(@"C:\Users\gabri\Desktop\ISKOLA\6.félév\C#\egyetemikurzus-2022\toplist.json"))
+                {
+                    string json = r.ReadToEnd();
+                    players = JsonSerializer.Deserialize<List<Player>>(json);
+                }
+            } else
+            {
+                List<Player> playerList = new List<Player>();
+                string playersJson = JsonSerializer.Serialize(playerList);
+                File.WriteAllText(@"C:\Users\gabri\Desktop\ISKOLA\6.félév\C#\egyetemikurzus-2022\toplist.json", playersJson);
+                loadPlayersData();
             }
         }
 
@@ -64,7 +58,8 @@ namespace tic_tac_toe
             if (GameForm.p1 != null && GameForm.p2 != null)
             {
                 loadform(new GameForm());
-            } else
+            }
+            else
             {
                 MessageBox.Show("You are need to be logged in.", "TicTacToe",
                     MessageBoxButtons.OK, MessageBoxIcon.Information);
@@ -84,6 +79,11 @@ namespace tic_tac_toe
         private void login_btn_Click(object sender, EventArgs e)
         {
             loadform(new LoginForm());
+        }
+
+        private void MenuForm_Load(object sender, EventArgs e)
+        {
+
         }
     }
 }
